rename coretesting -> testutil
