+++ collections/collections_test.go
@@ -7,14 +7,14 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"cosmossdk.io/collections/internal/coretesting"
+	"cosmossdk.io/collections/internal/testutil"
 
 	store "cosmossdk.io/collections/corecompat"
 )
 
 func deps() (store.KVStoreService, context.Context) {
-	ctx := coretesting.Context()
-	kv := coretesting.KVStoreService(ctx, "test")
+	ctx := testutil.Context()
+	kv := testutil.KVStoreService(ctx, "test")
 	return kv, ctx
 }
