+++ collections/indexes/indexes_test.go
@@ -3,14 +3,14 @@ package indexes
 import (
 	"context"
 
-	"cosmossdk.io/collections/internal/coretesting"
+	"cosmossdk.io/collections/internal/testutil"
 
 	store "cosmossdk.io/collections/corecompat"
 )
 
 func deps() (store.KVStoreService, context.Context) {
-	ctx := coretesting.Context()
-	kv := coretesting.KVStoreService(ctx, "test")
+	ctx := testutil.Context()
+	kv := testutil.KVStoreService(ctx, "test")
 	return kv, ctx
 }
